Completed setup of GNU/Global and Ctags.
In this commit I've fixed projectile to integrate correctly with helm.
Now helm actions can be performed in projectile-switch-project buffers.

GNU/Global is configured to point to the gtags section at the
~/.globalrc configuration file. Different projects can use a local
gtags.conf file.

To enable ggtags to read the correct configuration to generate the tags
for a project, set the `ggtags-process-environment` variable in
.dir-locals.el with the usual gtags environment variables indicating the
path to the config file and the section to read.

(ggtags-process-environment
 "GTAGSCONF=/Users/andreaturso/.globalrc"
 "GTAGSLABEL=justpark")

Note about GNU/Global and Ctags integration: the language names
in the gtags configuration file appear to be case sensitive.
I took me a long time to debug persistently empty GTAGS files,
while gtags -d provided only the empty index headers.